requirements.txt
asgiref==3.8.1
cloudinary==1.40.0
dj-database-url==0.5.0
Django==4.2.13
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
djangorestframework==3.15.1
Pillow==8.2.0
psycopg2==2.9.9
sqlparse==0.5.0
